Moved the Texture struct to it's own module.
